:hammer: fix migrations

+++ database/migrations/2026_05_26_000003_create_notifications_table.php
@@ -10,8 +10,8 @@ return new class extends Migration
     {
         Schema::create('notifications', function (Blueprint $table) {
             $table->id();
-            $table->unsignedBigInteger('user_id');
-            $table->unsignedBigInteger('santri_id')->nullable();
+            $table->unsignedInteger('user_id');
+            $table->unsignedInteger('santri_id')->nullable();
             $table->enum('type', ['pendaftaran', 'pembayaran', 'verifikasi', 'diterima', 'ditolak', 'dokumen']);
             $table->string('title');
             $table->text('message');
